test(logic): cover codec devices and source VFS
