+++ x/logic/fs/codec/fs_test.go
@@ -2,7 +2,6 @@ package codec
 
 import (
 	"bytes"
-	"encoding/json"
 	"errors"
 	"io"
 	"io/fs"
@@ -37,6 +36,7 @@ func TestAll(t *testing.T) {
 	Convey("Given the codec registry", t, func() {
 		So(slices.Contains(All(), "bech32"), ShouldBeTrue)
 		So(slices.Contains(All(), "json"), ShouldBeTrue)
+		So(slices.Contains(All(), "text"), ShouldBeTrue)
 	})
 }
 
@@ -462,6 +462,86 @@ func TestCodecDeviceFSFunctional(t *testing.T) {
 			request:        []byte("unknown\n{}"),
 			expectedOutput: "error(invalid_request).\n",
 		},
+
+		// Text codec tests
+		{
+			name:           "text encode utf8",
+			codecName:      "text",
+			request:        []byte("encode\ntext(utf8, aap)."),
+			expectedOutput: "ok([97,97,112]).\n",
+		},
+		{
+			name:           "text decode utf8",
+			codecName:      "text",
+			request:        []byte("decode\nbytes(utf8, [97,97,112])."),
+			expectedOutput: "ok([a,a,p]).\n",
+		},
+		{
+			name:           "text encode text",
+			codecName:      "text",
+			request:        []byte("encode\ntext(text, 'ù')."),
+			expectedOutput: "ok([195,185]).\n",
+		},
+		{
+			name:           "text decode text",
+			codecName:      "text",
+			request:        []byte("decode\nbytes(text, [195,185])."),
+			expectedOutput: "ok([ù]).\n",
+		},
+		{
+			name:           "text encode octet",
+			codecName:      "text",
+			request:        []byte("encode\ntext(octet, 'ù')."),
+			expectedOutput: "ok([249]).\n",
+		},
+		{
+			name:           "text decode octet",
+			codecName:      "text",
+			request:        []byte("decode\nbytes(octet, [249])."),
+			expectedOutput: "ok([ù]).\n",
+		},
+		{
+			name:           "text encode utf-16le",
+			codecName:      "text",
+			request:        []byte("encode\ntext('utf-16le', '今日は')."),
+			expectedOutput: "ok([202,78,229,101,111,48]).\n",
+		},
+		{
+			name:           "text decode utf-16be",
+			codecName:      "text",
+			request:        []byte("decode\nbytes('utf-16be', [0,97,0,97,0,112])."),
+			expectedOutput: "ok([a,a,p]).\n",
+		},
+		{
+			name:           "text encode invalid charset",
+			codecName:      "text",
+			request:        []byte("encode\ntext(foo, aap)."),
+			expectedOutput: "error(type_error(charset,foo)).\n",
+		},
+		{
+			name:           "text decode invalid byte",
+			codecName:      "text",
+			request:        []byte("decode\nbytes(latin2, [400])."),
+			expectedOutput: "error(type_error(byte,400)).\n",
+		},
+		{
+			name:           "text encode malformed term",
+			codecName:      "text",
+			request:        []byte("encode\ntext(utf8, aap). trailing"),
+			expectedOutput: "error(syntax_error(prolog(malformed_term))).\n",
+		},
+		{
+			name:           "text encode invalid request term",
+			codecName:      "text",
+			request:        []byte("encode\nfoo(utf8, aap)."),
+			expectedOutput: "error(invalid_request).\n",
+		},
+		{
+			name:           "text decode invalid bytes term",
+			codecName:      "text",
+			request:        []byte("decode\nbytes(utf8, foo(bar))."),
+			expectedOutput: "error(type_error(list,foo(bar))).\n",
+		},
 	}
 
 	Convey("Given a codec device filesystem", t, func() {
@@ -472,219 +552,6 @@ func TestCodecDeviceFSFunctional(t *testing.T) {
 				So(string(response), ShouldEqual, tc.expectedOutput)
 			})
 		}
-	})
-}
-
-func TestJSONCodecDecode(t *testing.T) {
-	testCases := []struct {
-		name     string
-		payload  []byte
-		expected string
-	}{
-		{
-			name:     "array with primitive values",
-			payload:  []byte(`[1,"two",false,null]`),
-			expected: "ok([1.0,two,@(false),@(null)]).\n",
-		},
-		{
-			name:     "boolean true",
-			payload:  []byte(`true`),
-			expected: "ok(@(true)).\n",
-		},
-		{
-			name:     "number",
-			payload:  []byte(`1.5`),
-			expected: "ok(1.5).\n",
-		},
-		{
-			name:     "unexpected EOF",
-			payload:  []byte(`{"foo":`),
-			expected: "error(syntax_error(json(eof))).\n",
-		},
-	}
-
-	Convey("Given a JSON codec decoder", t, func() {
-		codec := &jsonCodec{}
-		for _, tc := range testCases {
-			Convey(tc.name, func() {
-				So(renderTerm(t, codec.Decode(tc.payload)), ShouldEqual, tc.expected)
-			})
-		}
-	})
-}
-
-func TestJSONCodecEncode(t *testing.T) {
-	testCases := []struct {
-		name     string
-		payload  []byte
-		expected string
-	}{
-		{
-			name:     "atom",
-			payload:  []byte("hello."),
-			expected: "ok('\"hello\"').\n",
-		},
-		{
-			name:     "empty list",
-			payload:  []byte("[]."),
-			expected: "ok([]).\n",
-		},
-		{
-			name:     "integer",
-			payload:  []byte("42."),
-			expected: "ok('42').\n",
-		},
-		{
-			name:     "float",
-			payload:  []byte("1.5."),
-			expected: "ok('1.5').\n",
-		},
-		{
-			name:     "array",
-			payload:  []byte("[foo,42,@(false),@(null),[]]."),
-			expected: "ok('[\"foo\",42,false,null,[]]').\n",
-		},
-		{
-			name:     "malformed term",
-			payload:  []byte("."),
-			expected: "error(syntax_error(prolog(malformed_term))).\n",
-		},
-	}
-
-	Convey("Given a JSON codec encoder", t, func() {
-		codec := &jsonCodec{}
-		for _, tc := range testCases {
-			Convey(tc.name, func() {
-				So(renderTerm(t, codec.Encode(tc.payload)), ShouldEqual, tc.expected)
-			})
-		}
-	})
-}
-
-func TestJSONCodecEncodeWriterErrors(t *testing.T) {
-	testCases := []struct {
-		name       string
-		termSource string
-		failAt     int
-	}{
-		{
-			name:       "object opening",
-			termSource: "json([foo=bar]).",
-			failAt:     0,
-		},
-		{
-			name:       "object key",
-			termSource: "json([foo=bar]).",
-			failAt:     1,
-		},
-		{
-			name:       "object separator",
-			termSource: "json([foo=bar]).",
-			failAt:     2,
-		},
-		{
-			name:       "object value",
-			termSource: "json([foo=bar]).",
-			failAt:     3,
-		},
-		{
-			name:       "object comma",
-			termSource: "json([foo=bar,baz=qux]).",
-			failAt:     4,
-		},
-		{
-			name:       "object closing",
-			termSource: "json([foo=bar]).",
-			failAt:     4,
-		},
-		{
-			name:       "array opening",
-			termSource: "[foo].",
-			failAt:     0,
-		},
-		{
-			name:       "array value",
-			termSource: "[foo].",
-			failAt:     1,
-		},
-		{
-			name:       "array comma",
-			termSource: "[foo,bar].",
-			failAt:     2,
-		},
-		{
-			name:       "array closing",
-			termSource: "[foo].",
-			failAt:     2,
-		},
-	}
-
-	Convey("Given JSON encoder writer failures", t, func() {
-		for _, tc := range testCases {
-			Convey(tc.name, func() {
-				term, err := parseJSONTerm([]byte(tc.termSource))
-				So(err, ShouldBeNil)
-
-				err = encodeTermToJSON(term, &failAtWriter{failAt: tc.failAt, err: io.ErrClosedPipe}, engine.NewEnv())
-				So(err, ShouldEqual, io.ErrClosedPipe)
-			})
-		}
-	})
-}
-
-func TestJSONCodecErrorTerms(t *testing.T) {
-	testCases := []struct {
-		name     string
-		err      error
-		expected string
-	}{
-		{
-			name:     "unmarshal type",
-			err:      &json.UnmarshalTypeError{Offset: 7, Value: "number"},
-			expected: "syntax_error(json(malformed_json(7,number))).\n",
-		},
-		{
-			name:     "EOF",
-			err:      io.EOF,
-			expected: "syntax_error(json(eof)).\n",
-		},
-		{
-			name:     "unknown",
-			err:      errors.New("boom"),
-			expected: "syntax_error(json(unknown)).\n",
-		},
-	}
-
-	Convey("Given JSON error term mapping", t, func() {
-		for _, tc := range testCases {
-			Convey(tc.name, func() {
-				So(renderTerm(t, jsonErrorTerm(tc.err)), ShouldEqual, tc.expected)
-			})
-		}
-
-		Convey("when an encoder error is not a Prolog exception", func() {
-			So(renderTerm(t, exceptionFormal(errors.New("boom"))), ShouldEqual, "system_error.\n")
-		})
-	})
-}
-
-func TestJSONMarshalToStreamErrors(t *testing.T) {
-	Convey("Given JSON marshaling to a stream", t, func() {
-		Convey("when Go JSON rejects the value", func() {
-			var buf bytes.Buffer
-			err := marshalToJSONStream(math.Inf(1), engine.NewAtom("inf"), &buf, engine.NewEnv())
-			So(err, ShouldNotBeNil)
-		})
-	})
-}
-
-func TestJSONEncodeTermValidation(t *testing.T) {
-	Convey("Given JSON term validation", t, func() {
-		Convey("when an unsupported term implementation is encoded", func() {
-			var buf bytes.Buffer
-			err := encodeTermToJSON(unknownTerm{}, &buf, engine.NewEnv())
-			So(renderTerm(t, exceptionFormal(err)), ShouldEqual, "type_error(json,unknown).\n")
-		})
 	})
 }
 
@@ -870,21 +737,6 @@ func (w errWriter) Write(_ []byte) (int, error) {
 	return 0, w.err
 }
 
-type failAtWriter struct {
-	writes int
-	failAt int
-	err    error
-}
-
-func (w *failAtWriter) Write(p []byte) (int, error) {
-	if w.writes == w.failAt {
-		return 0, w.err
-	}
-
-	w.writes++
-	return len(p), nil
-}
-
 type badTerm struct {
 	err error
 }
@@ -894,17 +746,6 @@ func (t badTerm) WriteTerm(_ io.Writer, _ *engine.WriteOptions, _ *engine.Env) e
 }
 
 func (t badTerm) Compare(_ engine.Term, _ *engine.Env) int {
-	return 0
-}
-
-type unknownTerm struct{}
-
-func (t unknownTerm) WriteTerm(w io.Writer, _ *engine.WriteOptions, _ *engine.Env) error {
-	_, err := w.Write([]byte("unknown"))
-	return err
-}
-
-func (t unknownTerm) Compare(_ engine.Term, _ *engine.Env) int {
 	return 0
 }
